+++ common/reviews/api/api-extractor.api.ts
@@ -49,6 +49,7 @@ class ApiJsonGenerator extends ApiItemVisitor {
 }
 
 class ExternalApiHelper {
+  // (undocumented)
   public static generateApiJson(rootDir: string, libFolder: string, externalPackageFilePath: string): void;
 }